requirements.txt
python_dotenv

langchain
langchain_community
langchain_openai
langchain_core
langchain_groq

streamlit